app/src/auction/serializers/profile_serializers.py
from auction.models import Profile
from rest_framework import serializers


class ProfileDetailSerializer(serializers.ModelSerializer):
    user = serializers.HiddenField(default=serializers.CurrentUserDefault())

    class Meta:
        model = Profile
        fields = '__all__'
