app/requirements.txt
Django==3.1
djangorestframework==3.11.1
djoser==2.0.3
psycopg2-binary==2.8.5